chore(shared): rename to variables
