+++ src/Components/Index/IndexScene.tsx
@@ -26,7 +26,14 @@ import {
 import { Text, useStyles2 } from '@grafana/ui';
 import { ServiceScene } from 'Components/Service/ServiceScene';
 import { Pattern } from 'Components/Tabs/Patterns/Pattern';
-import { VAR_DATASOURCE, VAR_FIELDS, VAR_FILTERS, VAR_LINE_FILTER, VAR_PATTERNS, explorationDS } from 'services/shared';
+import {
+  VAR_DATASOURCE,
+  VAR_FIELDS,
+  VAR_FILTERS,
+  VAR_LINE_FILTER,
+  VAR_PATTERNS,
+  explorationDS,
+} from 'services/variables';
 import { ServiceSelectionComponent } from './ServiceSelectionComponent';
 
 import pluginJson from '../../plugin.json';
