+++ src/Components/Tabs/Labels/LabelBreakdownScene.tsx
@@ -34,7 +34,7 @@ import {
   ALL_VARIABLE_VALUE,
   explorationDS,
   LOG_STREAM_SELECTOR_EXPR,
-} from 'services/shared';
+} from 'services/variables';
 import { getLokiDatasource, getLabelOptions } from 'services/scenes';
 import { FieldSelector } from 'Components/Forms/FieldSelector';
